requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openai-whisper==20231117
torch==2.3.1
ffmpeg-python==0.2.0
yt-dlp>=2024.08.06
aiohttp==3.9.1
pydub==0.25.1
reportlab==4.0.4
Pillow==10.0.0
Flask>=2.3.0
gunicorn>=21.2.0
